requirements.txt
aiosqlite==0.16.1
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cmake==3.18.4.post1
colorama==0.4.4
databases==0.4.1
dlib==19.21.1
flake8==3.8.4
Flask==1.1.2
googledrivedownloader==0.4
idna==2.10
imutils==0.5.2
iniconfig==1.1.1
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
makefun==1.9.5
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.800
mypy-extensions==0.4.3
numpy==1.19.5
opencv-python==4.5.1.48
packaging==20.8
Pillow==8.1.0
pluggy==0.13.1
progress==1.5
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.6.0
pyparsing==2.4.7
pyTelegramBotAPI==3.7.6
pytest==6.2.2
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
transliterate==1.10.2
typed-ast==1.4.2
typesystem==0.2.4
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
